edl_seek_type is used as CONF_TYPE_FLAG, which is a 32bit-Type (fixes possible memory corruption).
git-svn-id: svn://git.mplayerhq.hu/mplayer/trunk@34366 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ mencoder.c
@@ -238,7 +238,7 @@ static edl_record_ptr edl_records = NULL; ///< EDL entries memory area
 static edl_record_ptr next_edl_record = NULL; ///< only for traversing edl_records
 static short edl_muted; ///< Stores whether EDL is currently in muted mode.
 static short edl_seeking; ///< When non-zero, stream is seekable.
-static short edl_seek_type; ///< When non-zero, frames are discarded instead of seeking.
+static int edl_seek_type; ///< When non-zero, frames are discarded instead of seeking.
 
 /* This header requires all the global variable declarations. */
 #include "cfg-mencoder.h"
